+++ src/views/SystemWorkspace/index.tsx
@@ -15,6 +15,7 @@ import MainContent from 'components/MainContent';
 import PageHeading from 'components/PageHeading';
 import PageLoading from 'components/PageLoading';
 import { getAtoStatus } from 'components/shared/AtoStatus';
+import Breadcrumbs from 'components/shared/Breadcrumbs';
 import SystemProfileModal from 'components/SystemProfileModal';
 import TLCTag from 'components/TLCTag';
 import GetSystemWorkspaceQuery from 'queries/GetSystemWorkspaceQuery';
@@ -25,7 +26,6 @@ import {
 import { RoleTypeName } from 'types/systemProfile';
 import linkCedarSystemIdQueryString from 'utils/linkCedarSystemIdQueryString';
 import NotFound from 'views/NotFound';
-import Breadcrumbs from 'views/TechnicalAssistance/Breadcrumbs';
 
 import AtoCard from './components/AtoCard';
 import HelpLinks from './components/HelpLinks';
